Addressed Shreay and Sourav's comments.

+++ infrastructure/src/main/java/org/corfudb/infrastructure/ServerContext.java
@@ -290,7 +290,7 @@ public class ServerContext implements AutoCloseable {
      */
     public long getLogReplicationMaxDataMessageSize() {
         String val = getServerConfig(String.class, "--max-replication-data-message-size");
-        return val == null ? DEFAULT_MAX_DATA_MSG_SIZE : Integer.parseInt(val);
+        return val == null ? DEFAULT_MAX_DATA_MSG_SIZE : Long.parseLong(val);
     }
 
     /**
